stricter ints for round and valIndex (#4582)
* ints: change some ints to int32

- this is to help reduce the overall size of some proto prs.


* add safemath for int32

* add safemath for subtraction

* add go doc to safemath funcs

* move panics to safe math, add checkconvert func

* make safe convert accept int64

+++ consensus/byzantine_test.go
@@ -56,12 +56,12 @@ func TestByzantine(t *testing.T) {
 			// NOTE: Now, test validators are MockPV, which by default doesn't
 			// do any safety checks.
 			css[i].privValidator.(types.MockPV).DisableChecks()
-			css[i].decideProposal = func(j int) func(int64, int) {
-				return func(height int64, round int) {
+			css[i].decideProposal = func(j int32) func(int64, int32) {
+				return func(height int64, round int32) {
 					byzantineDecideProposalFunc(t, height, round, css[j], switches[j])
 				}
-			}(i)
-			css[i].doPrevote = func(height int64, round int) {}
+			}(int32(i))
+			css[i].doPrevote = func(height int64, round int32) {}
 		}
 
 		eventBus := css[i].eventBus
@@ -172,7 +172,7 @@ func TestByzantine(t *testing.T) {
 //-------------------------------
 // byzantine consensus functions
 
-func byzantineDecideProposalFunc(t *testing.T, height int64, round int, cs *State, sw *p2p.Switch) {
+func byzantineDecideProposalFunc(t *testing.T, height int64, round int32, cs *State, sw *p2p.Switch) {
 	// byzantine user should create two proposals and try to split the vote.
 	// Avoid sending on internalMsgQueue and running consensus state.
 
@@ -209,7 +209,7 @@ func byzantineDecideProposalFunc(t *testing.T, height int64, round int, cs *Stat
 
 func sendProposalAndParts(
 	height int64,
-	round int,
+	round int32,
 	cs *State,
 	peer p2p.Peer,
 	proposal *types.Proposal,
