+++ types/vote_set_test.go
@@ -15,7 +15,7 @@ import (
 // NOTE: privValidators are in order
 func randVoteSet(
 	height int64,
-	round int,
+	round int32,
 	signedMsgType SignedMsgType,
 	numValidators int,
 	votingPower int64,
@@ -25,7 +25,7 @@ func randVoteSet(
 }
 
 // Convenience: Return new vote with different validator address/index
-func withValidator(vote *Vote, addr []byte, idx int) *Vote {
+func withValidator(vote *Vote, addr []byte, idx int32) *Vote {
 	vote = vote.Copy()
 	vote.ValidatorAddress = addr
 	vote.ValidatorIndex = idx
@@ -40,7 +40,7 @@ func withHeight(vote *Vote, height int64) *Vote {
 }
 
 // Convenience: Return new vote with different round
-func withRound(vote *Vote, round int) *Vote {
+func withRound(vote *Vote, round int32) *Vote {
 	vote = vote.Copy()
 	vote.Round = round
 	return vote
@@ -68,7 +68,7 @@ func withBlockPartsHeader(vote *Vote, blockPartsHeader PartSetHeader) *Vote {
 }
 
 func TestAddVote(t *testing.T) {
-	height, round := int64(1), 0
+	height, round := int64(1), int32(0)
 	voteSet, _, privValidators := randVoteSet(height, round, PrevoteType, 10, 1)
 	val0 := privValidators[0]
 
@@ -116,7 +116,7 @@ func TestAddVote(t *testing.T) {
 }
 
 func Test2_3Majority(t *testing.T) {
-	height, round := int64(1), 0
+	height, round := int64(1), int32(0)
 	voteSet, _, privValidators := randVoteSet(height, round, PrevoteType, 10, 1)
 
 	voteProto := &Vote{
@@ -129,7 +129,7 @@ func Test2_3Majority(t *testing.T) {
 		BlockID:          BlockID{nil, PartSetHeader{}},
 	}
 	// 6 out of 10 voted for nil.
-	for i := 0; i < 6; i++ {
+	for i := int32(0); i < 6; i++ {
 		pubKey, err := privValidators[i].GetPubKey()
 		require.NoError(t, err)
 		addr := pubKey.Address()
@@ -178,7 +178,7 @@ func Test2_3Majority(t *testing.T) {
 }
 
 func Test2_3MajorityRedux(t *testing.T) {
-	height, round := int64(1), 0
+	height, round := int64(1), int32(0)
 	voteSet, _, privValidators := randVoteSet(height, round, PrevoteType, 100, 1)
 
 	blockHash := crypto.CRandBytes(32)
@@ -196,7 +196,7 @@ func Test2_3MajorityRedux(t *testing.T) {
 	}
 
 	// 66 out of 100 voted for nil.
-	for i := 0; i < 66; i++ {
+	for i := int32(0); i < 66; i++ {
 		pubKey, err := privValidators[i].GetPubKey()
 		require.NoError(t, err)
 		addr := pubKey.Address()
@@ -295,7 +295,7 @@ func Test2_3MajorityRedux(t *testing.T) {
 }
 
 func TestBadVotes(t *testing.T) {
-	height, round := int64(1), 0
+	height, round := int64(1), int32(0)
 	voteSet, _, privValidators := randVoteSet(height, round, PrevoteType, 10, 1)
 
 	voteProto := &Vote{
@@ -370,7 +370,7 @@ func TestBadVotes(t *testing.T) {
 }
 
 func TestConflicts(t *testing.T) {
-	height, round := int64(1), 0
+	height, round := int64(1), int32(0)
 	voteSet, _, privValidators := randVoteSet(height, round, PrevoteType, 4, 1)
 	blockHash1 := tmrand.Bytes(32)
 	blockHash2 := tmrand.Bytes(32)
@@ -513,7 +513,7 @@ func TestConflicts(t *testing.T) {
 }
 
 func TestMakeCommit(t *testing.T) {
-	height, round := int64(1), 0
+	height, round := int64(1), int32(0)
 	voteSet, _, privValidators := randVoteSet(height, round, PrecommitType, 10, 1)
 	blockHash, blockPartsHeader := crypto.CRandBytes(32), PartSetHeader{123, crypto.CRandBytes(32)}
 
@@ -528,7 +528,7 @@ func TestMakeCommit(t *testing.T) {
 	}
 
 	// 6 out of 10 voted for some block.
-	for i := 0; i < 6; i++ {
+	for i := int32(0); i < 6; i++ {
 		pv, err := privValidators[i].GetPubKey()
 		assert.NoError(t, err)
 		addr := pv.Address()
